refactor: unify native folder name to capacitor
