+++ apps/web/src/shared/lib/capacitor/app-version/compareVersion.ts
