+++ apps/web/src/shared/lib/capacitor/app-version/fetchMinVersion.ts
