+++ apps/web/src/shared/lib/capacitor/app-version/forceUpdate.ts
