+++ apps/web/src/shared/lib/capacitor/app-version/getAppVersion.ts
