+++ apps/web/src/shared/lib/capacitor/app-version/index.ts
