+++ apps/web/src/shared/lib/capacitor/app-version/openAppStore.ts
