+++ apps/web/src/shared/lib/capacitor/auth/index.ts
