+++ apps/web/src/shared/lib/capacitor/index.ts
